# app/backend/openapi/openapi.v1.yaml
openapi: 3.0.3
info:
  title: AeroGenIoT API
  version: 1.0.0
  description: API base para AeroGenIoT (monitor de cultivos)
servers:
  - url: /api/v1
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Registro con email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          $ref: "#/components/responses/OkAuth"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /auth/login:
    post:
      tags: [auth]
      summary: Login con email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          $ref: "#/components/responses/OkAuth"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /auth/google:
    post:
      tags: [auth]
      summary: Login con Google (Gmail)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleLoginRequest"
      responses:
        "200":
          $ref: "#/components/responses/OkAuth"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotacion de refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          $ref: "#/components/responses/OkAuth"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /auth/logout:
    post:
      tags: [auth]
      summary: Cierra sesion y revoca refresh token
      security:
        - bearerAuth: []
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /auth/me:
    get:
      tags: [auth]
      summary: Perfil del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          headers:
            x-request-id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /farms:
    get:
      tags: [farms]
      summary: Lista granjas del usuario
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags: [farms]
      summary: Crea una granja
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /farms/{farmId}:
    get:
      tags: [farms]
      summary: Obtiene una granja
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FarmId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [farms]
      summary: Actualiza una granja
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FarmId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [farms]
      summary: Elimina una granja
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FarmId"
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /farms/{farmId}/plots:
    get:
      tags: [plots]
      summary: Lista plots de una granja
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FarmId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plot"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [plots]
      summary: Crea un plot
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FarmId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlotCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /plots/{plotId}:
    get:
      tags: [plots]
      summary: Obtiene un plot
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PlotId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plot"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [plots]
      summary: Actualiza un plot
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PlotId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlotUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [plots]
      summary: Elimina un plot
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PlotId"
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /devices:
    get:
      tags: [devices]
      summary: Lista dispositivos
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [devices]
      summary: Registra un dispositivo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /devices/{deviceId}:
    get:
      tags: [devices]
      summary: Obtiene un dispositivo
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [devices]
      summary: Actualiza un dispositivo
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /devices/{deviceId}/assign:
    post:
      tags: [devices]
      summary: Asigna dispositivo a granja/plot
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceAssign"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /readings:
    get:
      tags: [readings]
      summary: Lista lecturas
      security:
        - bearerAuth: []
      parameters:
        - name: farmId
          in: query
          schema:
            type: string
        - name: plotId
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SensorReading"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /readings/latest:
    get:
      tags: [readings]
      summary: Ultima lectura por granja/plot
      security:
        - bearerAuth: []
      parameters:
        - name: farmId
          in: query
          required: true
          schema:
            type: string
        - name: plotId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorReading"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /readings/ingest:
    post:
      tags: [readings]
      summary: Ingesta de lectura desde edge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestPayload"
      responses:
        "202":
          description: Accepted
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
  /recommendations:
    get:
      tags: [recommendations]
      summary: Lista recomendaciones
      security:
        - bearerAuth: []
      parameters:
        - name: farmId
          in: query
          schema:
            type: string
        - name: plotId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /recommendations/latest:
    get:
      tags: [recommendations]
      summary: Ultima recomendacion por granja/plot
      security:
        - bearerAuth: []
      parameters:
        - name: farmId
          in: query
          required: true
          schema:
            type: string
        - name: plotId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /notifications/register:
    post:
      tags: [notifications]
      summary: Registra token push (Expo)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPushRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushToken"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /notifications/test:
    post:
      tags: [notifications]
      summary: Envia notificacion de prueba
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    RequestId:
      description: Request id
      schema:
        type: string
  parameters:
    FarmId:
      name: farmId
      in: path
      required: true
      schema:
        type: string
    PlotId:
      name: plotId
      in: path
      required: true
      schema:
        type: string
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    OkAuth:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AuthResponse"
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Too Many Requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: validation_error
        message:
          type: string
          example: Invalid payload
        details:
          type: object
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
          example: user@AeroGenIoT.app
        password:
          type: string
          format: password
          example: "secret123"
        name:
          type: string
          example: "Juan"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    GoogleLoginRequest:
      type: object
      required: [idToken]
      properties:
        idToken:
          type: string
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, viewer]
...
    Farm:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        location:
          type: string
        createdAt:
          type: string
          format: date-time
    FarmCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        location:
          type: string
    FarmUpdate:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
    Plot:
      type: object
      properties:
        id:
          type: string
        farmId:
          type: string
        name:
          type: string
        cropType:
          type: string
    PlotCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        cropType:
          type: string
    PlotUpdate:
      type: object
      properties:
        name:
          type: string
        cropType:
          type: string
    Device:
      type: object
      properties:
        id:
          type: string
        farmId:
          type: string
        plotId:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [provisioned, active, offline, retired]
    DeviceCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
    DeviceUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [provisioned, active, offline, retired]
    DeviceAssign:
      type: object
      properties:
        farmId:
          type: string
        plotId:
          type: string
    SensorReading:
      type: object
      properties:
        id:
          type: string
        deviceId:
          type: string
        farmId:
          type: string
        plotId:
          type: string
        timestamp:
          type: string
          format: date-time
        humidity:
          type: number
        temperature:
          type: number
        airTemperature:
          type: number
        airHumidity:
          type: number
        airPressure:
          type: number
        windSpeed:
          type: number
        solarRadiation:
          type: number
        rainfall:
          type: number
        et0:
          type: number
        etc:
          type: number
        kc:
          type: number
        kcStage:
          type: string
        daysPlanted:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        ph:
          type: number
        ce:
          type: number
        n:
          type: number
        p:
          type: number
        k:
          type: number
        season:
          type: string
    Recommendation:
      type: object
      properties:
        id:
          type: string
        readingId:
          type: string
        cantidadAgua:
          type: number
        recomendarRiego:
          type: boolean
        motivo:
          type: string
...
    IngestPayload:
      type: object
      required: [timestamp, sensor_values]
      properties:
        timestamp:
          type: string
          format: date-time
        deviceId:
          type: string
        farmId:
          type: string
        plotId:
          type: string
        sensor_values:
          type: object
        decision:
          type: object
        mode:
          type: string
    RegisterPushRequest:
      type: object
      required: [token, platform]
      properties:
        token:
          type: string
        platform:
          type: string
          example: android
        deviceId:
          type: string
    PushToken:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        token:
          type: string
        platform:
          type: string
        deviceId:
          type: string
        createdAt:
          type: string
          format: date-time